Redo config so its less cluttered
